pyproject.toml
[project]
name = "briefly-bot"
version = "0.1.0"
description = "A slack bot that fetches latest newses and trends in the Ai field and creates a short read to be published in a slack channel"
authors = [
    {name = "Dheeraj-Nalapat",email = "dheerajnalapat3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # general dependencies
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",

    #beautifulsoup4 PoC
    "beautifulsoup4>=4.12.0",

    #feedparser PoC
    "feedparser>=6.0.0",

    #scrapy PoC
    "scrapy>=2.10.0",

    #news through api PoC
    "arxiv>=2.2.0",
    "newsapi-python>=0.2.6",

    # news explorer PoC
    "trafilatura>=1.8.0",
    "tavily-python>=0.3.1",

    # Slack bot dependencies
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    
    # LLM dependencies (for summarization)
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-openai (>=0.3.30,<0.4.0)",
]

[tool.poetry.group.dev.dependencies]
autoflake = "2.3.1"
black = "25.1.0"
isort = "6.0.1"
pre-commit = "4.2.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint.messages_control]
disable = ["broad-except"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
